backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
django-ninja==1.3.0
django-silk==5.2.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2